+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## New Features
 
+- PR #3693 add string support, skipna to scan operation
 - PR #3662 Define and implement `shift`.
 - PR #3861 Added Series.sum feature for String